% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\docType{data}
\name{base_grid_bng}
\alias{base_grid_bng}
\title{Britain and Ireland TenKM grids in British National Grid}
\format{A shape file with 11100 rows and 1 variables
\describe{
  \item{tenkm}{polygon 10KM BNG grid square label}
}}
\usage{
base_grid_bng
}
\description{
A simple features data frame of the British and Irish 10KM 
grid squares in British National Grid projection (EPSG::27700)
}
\keyword{datasets}
