% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\docType{data}
\name{base_map}
\alias{base_map}
\title{Britain and Ireland Oil and Gas Authority's coastline}
\format{A shape file with 6 rows and 2 variables
\describe{
  \item{Country}{multipolygon country name}
  \item{UK}{multipolygon present in the UK or not (TRUE/FALSE)}
}}
\usage{
base_map
}
\description{
A simple features data frame of the British and Irish coastline 
derived from the Oil and Gas Authority's OGA and Lloyd's Register 
SNS Regional Geological Maps (Open Source) layer. 
Licence: Open Government Licence v3 (OGL) 
(http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/).
Attribution: Contains data © 2017 Oil & Gas Authority.
}
\keyword{datasets}
