% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{code_look_up}
\alias{code_look_up}
\title{Code changes}
\format{A list with 1 row and 2 variables
\describe{
  \item{code_old}{species or habitat code used in previous reporting rounds}
  \item{code_new}{species or habitat code  used in the latest 2019 reporting round}
}}
\usage{
code_look_up
}
\description{
A dataset containing the currently accepted species 
or habitat code and the old species or habitat code 
used in previous Article 17 reporting rounds
}
\keyword{datasets}
