% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{export_results_log}
\alias{export_results_log}
\title{Export results log}
\usage{
export_results_log(results, data_files, log_path)
}
\arguments{
\item{results}{list, contains the errors created from purrr's safely function}

\item{data_files}{vector, used to iterate through the safe function}

\item{log_path}{character or path object, path where log file will be exported}
}
\value{
csv, export log as a csv file
}
\description{
This function exports a log of success and failures resulting from
applying purrr's map functionality to iterate through a vector, applying
a safe function using purrr's safely functionality
}
