% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{extract_uk_assessment}
\alias{extract_uk_assessment}
\title{Extract UK assessment}
\usage{
extract_uk_assessment(data)
}
\arguments{
\item{data}{dataframe, created from read_aggregation_spreadsheet function}
}
\value{
dataframe, containing UK information spread across columns
}
\description{
This function extracts UK assessment from the data frame created by
using read_aggregation_spreadsheet
}
\examples{
extract_uk_assessment("general_information")
}
