% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_country}
\alias{format_country}
\title{Format country}
\usage{
format_country(country)
}
\arguments{
\item{country}{character, country name

 This function removes any spaces within the countries name}
}
\value{
character, country name without spaces
}
\description{
Format country
}
\examples{
format_country("Northen Ireland")
}
