% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_future_prospects}
\alias{format_future_prospects}
\title{Format future prospects}
\usage{
format_future_prospects(text)
}
\arguments{
\item{text}{character, UK future prospects criteria}
}
\value{
character, UK future prospectes criteria with
converted less than or equal sign
}
\description{
This function formats the uk future prospect fields
removing the less than or equal character.
}
\details{
R seems to convert the less than or equal character
to an equals character but exports as the original less than
or equal character. This is converted into unicode by SQL Server
Criteria also includes text one percent or less, so less than or 
equal character removed
}
\examples{
format_future_prospects("Negative - decreasing ≤1\% (one percent or less) per year on average")
}
