% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_habitat_code}
\alias{format_habitat_code}
\title{Format habitat code}
\usage{
format_habitat_code(habitatcode)
}
\arguments{
\item{habitatcode}{character, habitat code beginning with H}
}
\value{
character, habitat code without the initial H
}
\description{
UK habitat codes starts with the letter H to distinguish
them from species codes. European codes do not have this
initial letter and so this function removes the initial H
}
\examples{
format_habitat_code("H1234")
}
