% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_species_code}
\alias{format_species_code}
\title{Format species code}
\usage{
format_species_code(speciescode)
}
\arguments{
\item{speciescode}{character, species code beginning with S}
}
\value{
character, species code without the initial S
}
\description{
UK species codes starts with the letter S to distinguish
them from habitat codes. European codes do not have this
initial letter and so this function removes the initial S
}
\examples{
format_species_code("S1234")
}
