% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_sqlserver_connection}
\alias{get_sqlserver_connection}
\title{Get SQL Server connection}
\usage{
get_sqlserver_connection(keyring_name, service_name)
}
\arguments{
\item{keyring_name}{character, name of keyring}

\item{service_name}{chracter, name of service storing credentials}
}
\value{
connection object, SQL Server connection object
}
\description{
This function initiates a SQL Server connection
using the keyring package to store the database
credentials
}
\examples{
get_sqlserver_connection("keyring", "database_service_name")
}
