% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_sqlserver_connection_string}
\alias{get_sqlserver_connection_string}
\title{Get SQL Server connection string}
\usage{
get_sqlserver_connection_string(keyring_name, service_name,
  keyring = TRUE)
}
\arguments{
\item{keyring_name}{character, name of keyring}

\item{service_name}{character, name of service storing credentials}

\item{keyring}{logical, use of keyring TRUE (default) or FALSE}
}
\value{
character, connection string
}
\description{
This function returns the SQL Server connection string 
either using stored database credentials through the keyring 
package or else after requesting database credentials
}
\examples{
get_sqlserver_connection_string("keyring", "database_service_name")
}
