% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{graph_category_percentage}
\alias{graph_category_percentage}
\title{Graph of category percentages}
\usage{
graph_category_percentage(dataset, filter_column, x_column)
}
\arguments{
\item{dataset, }{dataframe, containing category data}

\item{filter_column}{character or factor, used for filtering}

\item{x_column}{character or factor, category data

This function is a shiny app which summarises the categories
percentages within a filtered selection in a plot. The function 
can be added to a markdown document using runtime shiny.}
}
\value{
shiny ggplot
}
\description{
Graph of category percentages
}
\examples{
graph_category_percentage("dataframe_name", "filter_column", "category_values")
}
