% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{information_habitat_look_up}
\alias{information_habitat_look_up}
\title{Habitat information}
\format{A tibble with 69 rows and 5 variables
\describe{
  \item{code}{habitat code starting with initial H}
  \item{name}{habitat name}
  \item{group}{habitat group}
  \item{alpha}{alpha values used when creating range alpha shape polygons}
  \item{clipping_polygon}{are the alpha shape range polygons clipped to the UK coast or to 10KM inline coastal grid squares}
}}
\usage{
information_habitat_look_up
}
\description{
A dataset containing the Article 17 habitat information
}
\keyword{datasets}
