% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{information_species_look_up}
\alias{information_species_look_up}
\title{Species information}
\format{A tibble with 76 rows and 10 variables
\describe{
  \item{code}{species code starting with initial S}
  \item{name}{scientific name}
  \item{common_name}{common name}
  \item{taxon_group}{taxonomic group}
  \item{occurrence}{is the species resident within the UK}
  \item{annex ii}{is the species on EC Habitat Directive's Annex II list}
  \item{annex iv}{is the species on EC Habitat Directive's Annex IV list}
  \item{annex v}{is the species on EC Habitat Directive's Annex V list}
  \item{alpha}{alpha values used when creating range alpha shape polygons}
  \item{clipping_polygon}{are the alpha shape range polygons clipped to the UK coast or to 10KM inline coastal grid squares}
}}
\usage{
information_species_look_up
}
\description{
A dataset containing the Article 17 species information
}
\keyword{datasets}
