% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{name_look_up}
\alias{name_look_up}
\title{Taxonomic name changes}
\format{A list with 1 row and 2 variables
\describe{
  \item{name_old}{scientific name used in previous reporting rounds}
  \item{name_new}{scientitic name used in the latest 2019 reporting round}
}}
\usage{
name_look_up
}
\description{
A dataset containing the currently accepted
scientific name and the old synonymous name used
in previous Article 17 reporting rounds
}
\keyword{datasets}
