% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{notes_habitat_look_up}
\alias{notes_habitat_look_up}
\title{Habitat audit notes categories}
\format{A tibble with 68 rows and 3 variables
\describe{
  \item{reporting_code}{habitat category's reporting code}
  \item{field_label}{habitat category's descriptive names, including reporting code}
  \item{entity_table_name}{reporting tool table name that contains column for the habitat category}
}}
\usage{
notes_habitat_look_up
}
\description{
A dataset containing the Article 17 habitat audit notes categories
}
\keyword{datasets}
