% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation.R
\name{read_aggregation_spreadsheet}
\alias{read_aggregation_spreadsheet}
\title{Read aggregation spreadsheet}
\usage{
read_aggregation_spreadsheet(spreadsheet_name, skip_rows = 7,
  spreadsheet_path = feature_path)
}
\arguments{
\item{spreadsheet_name}{character, name of spreadsheet to read}

\item{skip_rows}{integer, number of rows to skip at top of spreadsheet}

\item{spreadsheet_path}{character, path to spreadsheet to read}
}
\value{
dataframe, dataframe containing the three required columns
}
\description{
This function reads the specified aggregation spreadsheet, extracting
the UK information required for import into the UK reporting tool
}
\examples{
read_aggregation_spreadsheet("1 General Information")
}
