% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_assessments}
\alias{recode_assessments}
\title{Recode assessments}
\usage{
recode_assessments(assessments)
}
\arguments{
\item{assessments}{character, assessments text}
}
\value{
character, assessments abbreviated text
}
\description{
This function changes the descriptive assessments text
to the abbreviated text used in the reporting tool
}
\examples{
recode_assessments("Unfavourable-inadequate")
}
