% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_country_abbreviation}
\alias{recode_country_abbreviation}
\title{Recode country to abbrevation}
\usage{
recode_country_abbreviation(country)
}
\arguments{
\item{country}{character, country name}
}
\value{
character, country abbreviation
}
\description{
This function recodes the country to a corresponding
2 letter abbreviation. This is used for example in
the country file paths
}
\examples{
recode_country_abbreviation("N.Ireland")
}
