% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_country_agency}
\alias{recode_country_agency}
\title{Recode country to country agency}
\usage{
recode_country_agency(country)
}
\arguments{
\item{country}{character, country name}
}
\value{
character, country agency
}
\description{
This function recodes the country to the statutary
nature conservation body (country agency) within that 
country
}
\examples{
recode_country_agency("N.Ireland")
}
