% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_country_name}
\alias{recode_country_name}
\title{Recode country abbrevation to country name}
\usage{
recode_country_name(country_abbreviation)
}
\arguments{
\item{country_abbreviation}{character, country abbreviation}
}
\value{
character, country name
}
\description{
This function country abbreviation to the country
2name. This is used for example to convert country
abbreviations in file paths to the country name
}
\examples{
recode_country("NI")
}
