% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_favourable_range_operators}
\alias{recode_favourable_range_operators}
\title{Recode favourable range operators}
\usage{
recode_favourable_range_operators(operator)
}
\arguments{
\item{operator}{character, favourable range operators text}
}
\value{
character, abbreviated favourable range operators text
}
\description{
This function changes the descriptive favourable range operators 
text to the abbreviated text used in the reporting tool
}
\examples{
recode_favourable_range_operators("Much more than (>>)")
}
