% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_future_prospects_uk_fields}
\alias{recode_future_prospects_uk_fields}
\title{Recode future prospects UK fields}
\usage{
recode_future_prospects_uk_fields(prospect)
}
\arguments{
\item{prospect}{character, future prospects UK fields text}
}
\value{
character, future prospects UK fields text, less than or equal sign removed
}
\description{
This function changes the less than or equal
sign to ones that are handled in R
}
\examples{
recode_future_prospects_uk_fields ("Negative - decreasing ≤1\% (one percent or less) per year on average")
recode_future_prospects_uk_fields ("Very Negative - decreasing >1\% (more than one percent) per year on average")
recode_future_prospects_uk_fields ("Positive - increasing ≤1\% (one percent or less) per year on average")
recode_future_prospects_uk_fields ("Very Positive - increasing >1\% (more than one percent) per year on average")
}
