% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_measures_purpose_habitat}
\alias{recode_measures_purpose_habitat}
\title{Recode habitat measures purpose}
\usage{
recode_measures_purpose_habitat(purpose)
}
\arguments{
\item{purpose}{character, measures purpose text}
}
\value{
character, measures purpose abbreviated text
}
\description{
This function changes the descriptive measures purpose text
for habitat to the abbreviated text used in the reporting tool
}
\examples{
recode_measures_purpose_habitat("Maintain the current range, surface area or structure and functions of the habitat type")
}
