% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_prospects}
\alias{recode_prospects}
\title{Recode prospects}
\usage{
recode_prospects(prospect)
}
\arguments{
\item{prospect}{character, prospects text}
}
\value{
character, prospects abbreviated text
}
\description{
This function changes the descriptive prospects text
to the abbreviated text used in the reporting tool
}
\examples{
recode_prospects("Unknown")
}
