% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_ranking_type}
\alias{recode_ranking_type}
\title{Recode ranking type}
\usage{
recode_ranking_type(type)
}
\arguments{
\item{type}{character, ranking type text}
}
\value{
character, ranking type abbreviated text
}
\description{
This function changes the descriptive ranking type text
to the abbreviated text used in the reporting tool
}
\examples{
recode_ranking_type("pressure_ranking")
}
