% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_rate_of_decrease_uk_fields}
\alias{recode_rate_of_decrease_uk_fields}
\title{Recode rate of decrease UK fields}
\usage{
recode_rate_of_decrease_uk_fields(rate)
}
\arguments{
\item{rate}{character, rate of decrease UK fields text}
}
\value{
character, rate of decrease UK fields text, less than or equal sign removed
}
\description{
This function changes the less than or equal
sign to ones that are handled in R
}
\examples{
recode_rate_of_decrease_uk_fields ("Decreasing ≤1\% (one percent or less) per year on average")
recode_rate_of_decrease_uk_fields ("Decreasing >1\% (more than one percent) per year on average")
}
