% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_trends_conclusion}
\alias{recode_trends_conclusion}
\title{Recode trends conclusion}
\usage{
recode_trends_conclusion(trend)
}
\arguments{
\item{trend}{character, trends conclusion text}
}
\value{
character, trends conclusion abbreviation text
}
\description{
This function changes the descriptive trends conclusion text
to the abbreviated text used in the reporting tool
}
\examples{
recode_trends_conclusion("Improving (+)")
}
