% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_yes_no_unknown}
\alias{recode_yes_no_unknown}
\title{Recode yes, no or unknown}
\usage{
recode_yes_no_unknown(yes_no_unknown)
}
\arguments{
\item{yes_no_unknown}{character, yes, no or unknown text}
}
\value{
character, yes, no or unknown abbreviated text
}
\description{
This function changes the descriptive yes, no or unknown
text to the abbreviated text used in the reporting tool
}
\examples{
recode_yes_no_unknown("Unknown")
}
