% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{recode_yes_no_x}
\alias{recode_yes_no_x}
\title{Recode yes to x}
\usage{
recode_yes_no_x(yes_no_x)
}
\arguments{
\item{yes_no_x}{character, yes or no text}
}
\value{
character, x or NA_character_
}
\description{
This function changes yes to x used in the reporting tool
and removes no
}
\examples{
recode_yes_no_x("Yes")
}
