% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{use_test_data}
\alias{use_test_data}
\title{Use test data}
\usage{
use_test_data(data, test_data = TRUE)
}
\arguments{
\item{data}{data frame, all species or habitat data}

\item{test_data}{logical, whether to apply filter (default = TRUE) or not (FALSE)}
}
\value{
data frame, supplied or filter data for testing
}
\description{
This function filters data to selected species or
habitat codes. This is when testing the Article 17 
R scripts
}
\details{
Selected test species are S1317 Pipistrellus nathusii and S1013_Vertigo geyeri,
habitats are H6230 Species-rich Nardus grassland and H7130 Blanket bogs
}
\examples{
use_test_data(data)
}
