% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\docType{data}
\name{base_grid_etrs}
\alias{base_grid_etrs}
\title{TenKM grids in ETRS89 covering Britain and Ireland}
\format{A shape file with 28900 rows and 1 variables
\describe{
  \item{cell_code}{polygon 10KM ETRS grid square label}
}}
\usage{
base_grid_etrs
}
\description{
A shape file of 10KM grid squares covering Britain
and Ireland in ETRS89 projection (EPSG::3035)
}
\keyword{datasets}
