% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{format_future_prospects}
\alias{format_future_prospects}
\title{Format future prospects}
\usage{
format_future_prospects(text)
}
\arguments{
\item{text}{character, UK future prospects criteria}
}
\value{
character, UK future prospectes criteria with
converted less than or equal sign
}
\description{
This function formats the uk future prospect fields
converting the less than or equal sign to easier
handled characters.
}
\details{
It should not be used for criteria with equals in as
also converts equals to less than or equal
}
\examples{
format_future_prospects("Negative - decreasing ≤1\% (one percent or less) per year on average")
}
