% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\docType{data}
\name{notes_species_look_up}
\alias{notes_species_look_up}
\title{Species audit notes categories}
\format{A tibble with 80 rows and 3 variables
\describe{
  \item{reporting_code}{species category's reporting code}
  \item{field_label}{species category's descriptive names, including reporting code}
  \item{entity_table_name}{reporting tool table name that contains column for the species category}
}}
\usage{
notes_species_look_up
}
\description{
A dataset containing the Article 17 species audit notes categories
}
\keyword{datasets}
